wrap phrase splitter in own file, add to upload phrase data table too
